refactor(plugins): move React contexts into plugin dirs, delete PanelConfigs.tsx
SocialMediaConfigProvider/useSocialMediaConfig → plugins/socialSharing/context.ts
ImageCanvasConfigProvider/useImageCanvasConfig → plugins/imageCanvas/context.ts

Adds a TODO in ReactionCanvasAppV4 for making provider wiring generic
via a future PanelPlugin.provider field.
