+++ CHANGELOG.md
@@ -38,6 +38,7 @@ All notable changes to this project will be documented in this file. Releases cu
 - **Signature Canvas migrated to plugin** — `plugins/signatureCanvas/` is now a self-contained plugin; `strokeSegment` and `clearSignature` server routing removed from core `party/server.ts`; `signatureStrokes` state, `isPresenter` state, and all `activity === 'signature'` render blocks removed from `ReactionCanvasAppV4`; the two-mode UI (participant drawing vs. presenter grid) lives entirely inside the plugin component.
 - **Plugin API: `onRequest?` hook** — `ServerPlugin<S>` now supports an optional `onRequest?(request, ctx, state)` method; the main server delegates every HTTP request to plugins before its own switch, returning the first non-null `Response`; enables plugins that need REST endpoints (e.g. large-payload uploads).
 - **Steno + Story Tracer migrated to `narrative` plugin** — both panels moved into `plugins/narrative/` as two plugins sharing one folder (`[stenoPlugin, storyTracerPlugin]`); all Steno and StoryTracer state, message handlers, and the `storyTracerSetPoints` HTTP endpoint removed from core `party/server.ts`; `StoryTracerPoint`/`StoryTracerMeta` types moved into `plugins/narrative/types.ts`; `storyTracerUtils.ts` and `useEmbeddingWorker.ts` relocated into the plugin; `ActivityMode` entries and URL params (`?interface=steno`, `?interface=story-tracer`) are unchanged; lock release on disconnect moved into the steno plugin's `onClose` hook.
+- **Plugin-owned React contexts** — `SocialMediaConfigProvider`/`useSocialMediaConfig` moved into `plugins/socialSharing/context.ts`; `ImageCanvasConfigProvider`/`useImageCanvasConfig` moved into `plugins/imageCanvas/context.ts`; `app/context/PanelConfigs.tsx` deleted. A TODO in `ReactionCanvasAppV4` marks the remaining step of making provider wiring generic via a `PanelPlugin.provider` field.
 - **VoiceCall migrated to plugin** — `plugins/voiceCall/` is now a self-contained plugin; `callQueue`, `callPairs`, `callAlgorithm` state and all call-related message handlers removed from core `party/server.ts`; call-queue cleanup on disconnect moved into the plugin's `onClose` hook; `PluginContext` gains `sendToUser(userId, msg)` for targeted delivery (used by WebRTC signalling relay and hang-up notifications); `callAlgorithm` admin config modal now self-contained via `useAdminSocket()`; `WebRTCOfferEvent` / `HangUpCallEvent` etc. types moved from `party/types.ts` into `plugins/voiceCall/types.ts`. All plugin server tests consolidated to use `makeCtx` / `makeConn` from shared `plugins/testHelpers.ts`. ([feat/pluggable-panel-architecture](https://github.com/patcon/polislike-partykit-reaction-canvas/compare/main...feat/pluggable-panel-architecture))
 
 ### Fixed
