+++ plugins/socialSharing/component.stories.tsx
@@ -2,7 +2,7 @@ import type { Meta, StoryObj } from '@storybook/react-vite';
 import { expect, within } from 'storybook/test';
 import React from 'react';
 import SocialMediaPanel from './component';
-import { SocialMediaConfigProvider } from '../../app/context/PanelConfigs';
+import { SocialMediaConfigProvider } from './context';
 import type { SocialConfig } from '../../app/types';
 
 const EMPTY_CONFIG: SocialConfig = { default: '', twitter: '', bluesky: '', mastodon: '', instagram: '' };
